refactor: Resolve circular dependency causing TypeError
Refactored the placeholder registration logic to resolve a circular dependency between `placeholderManager.js` and `playerDataManager.js`.

- The `placeholderManager` no longer initializes default placeholders.
- The `registerPlaceholder` function is now exported.
- The `playerDataManager` now imports `registerPlaceholder` and registers its own `{topbal}` placeholder.

This change inverts the dependency, breaking the cycle and preventing a `TypeError: not a function` crash that occurred when `floatingTextManager` tried to resolve dynamic text.
