+++ AddonExeBP/scripts/core/playerDataManager.js
@@ -26,6 +26,7 @@ import { getConfig } from './configManager.js';
 import { world, system } from '@minecraft/server';
 import { debugLog, errorLog } from './logger.js';
 import { getPlayerFromCache } from './playerCache.js';
+import { registerPlaceholder } from './placeholderManager.js';
 
 const playerPropertyPrefix = 'exe:player.';
 const playerNameIdMapKey = 'exe:playerNameIdMap';
@@ -511,3 +512,28 @@ export function transfer(sourcePlayerId, targetPlayerId, amount) {
 
     return { success: true, message: 'Transfer successful.' };
 }
+
+registerPlaceholder('topbal', ({ index, valueKey }) => {
+    const leaderboard = getLeaderboard('balance');
+
+    // Validate leaderboard data and index
+    if (!Array.isArray(leaderboard) || index < 0 || index >= leaderboard.length) {
+        return '';
+    }
+
+    const playerData = leaderboard[index];
+
+    // Validate player data object
+    if (!playerData || typeof playerData !== 'object') {
+        return '';
+    }
+
+    if (valueKey === 'name') {
+        return playerData.name ?? ''; // Nullish coalescing for safety
+    }
+    if (valueKey === 'value') {
+        return String(playerData.balance ?? '0'); // Nullish coalescing for safety
+    }
+
+    return '';
+});
